refactor: 매장 식사 주문 핵심 하위 도메인 패키지 이동 및 불필요한 생성자 파라미터 제거
