+++ src/main/java/kitchenpos/order/tobe/eatinorder/domain/EatInOrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.eatinorder.domain;
+package kitchenpos.order.tobe.eatinorder.domain;
 
 public enum EatInOrderStatus {
     WAITING, ACCEPTED, SERVED, COMPLETED
