+++ src/main/java/kitchenpos/order/tobe/eatinorder/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.eatinorder.ordertable.domain;
+package kitchenpos.order.tobe.eatinorder.domain;
 
 import java.util.UUID;
 import javax.persistence.Column;
