+++ src/main/java/kitchenpos/order/tobe/eatinorder/infra/JpaEatInOrderRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.order.eatinorder.infra;
+package kitchenpos.order.tobe.eatinorder.infra;
 
 import java.util.UUID;
-import kitchenpos.order.eatinorder.domain.EatInOrder;
-import kitchenpos.order.eatinorder.domain.EatInOrderRepository;
+import kitchenpos.order.tobe.eatinorder.domain.EatInOrder;
+import kitchenpos.order.tobe.eatinorder.domain.EatInOrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaEatInOrderRepository extends EatInOrderRepository, JpaRepository<EatInOrder, UUID> {
