+++ src/main/java/kitchenpos/order/tobe/eatinorder/infra/JpaOrderTableRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.order.eatinorder.ordertable.infra;
+package kitchenpos.order.tobe.eatinorder.infra;
 
 import java.util.UUID;
-import kitchenpos.order.eatinorder.ordertable.domain.OrderTable;
-import kitchenpos.order.eatinorder.ordertable.domain.OrderTableRepository;
+import kitchenpos.order.tobe.eatinorder.domain.OrderTable;
+import kitchenpos.order.tobe.eatinorder.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
